chore: migrate functions utils to ESM
