+++ src/commands/functions/functions-build.mjs
@@ -1,9 +1,10 @@
 // @ts-check
 import { mkdir } from 'fs/promises'
 
+import { getFunctionsDir } from '../../utils/functions/index.mjs'
 import utils from '../../utils/index.cjs'
 
-const { NETLIFYDEVERR, NETLIFYDEVLOG, exit, getFunctionsDir, log } = utils
+const { NETLIFYDEVERR, NETLIFYDEVLOG, exit, log } = utils
 
 /**
  * The functions:build command
