+++ src/lib/functions/server.mjs
@@ -2,24 +2,17 @@
 import { get } from 'dot-prop'
 import jwtDecode from 'jwt-decode'
 
+import { CLOCKWORK_USERAGENT, getInternalFunctionsDir } from '../../utils/functions/index.mjs'
 import utils from '../../utils/index.cjs'
 
 import { handleBackgroundFunction, handleBackgroundFunctionResult } from './background.cjs'
-import { createFormSubmissionHandler } from './form-submissions-handler.cjs'
+import { createFormSubmissionHandler } from './form-submissions-handler.mjs'
 import { FunctionsRegistry } from './registry.cjs'
-import { handleScheduledFunction } from './scheduled.cjs'
+import { handleScheduledFunction } from './scheduled.mjs'
 import { handleSynchronousFunction } from './synchronous.cjs'
 import { shouldBase64Encode } from './utils.cjs'
 
-const {
-  CLOCKWORK_USERAGENT,
-  NETLIFYDEVERR,
-  NETLIFYDEVLOG,
-  error: errorExit,
-  generateNetlifyGraphJWT,
-  getInternalFunctionsDir,
-  log,
-} = utils
+const { NETLIFYDEVERR, NETLIFYDEVLOG, error: errorExit, generateNetlifyGraphJWT, log } = utils
 
 const buildClientContext = function (headers) {
   // inject a client context based on auth header, ported over from netlify-lambda (https://github.com/netlify/netlify-lambda/pull/57)
