--- src/utils/functions/constants.cjs
@@ -1,5 +1,0 @@
-const CLOCKWORK_USERAGENT = 'Netlify Clockwork'
-
-module.exports = {
-  CLOCKWORK_USERAGENT,
-}
