+++ src/utils/functions/constants.mjs
@@ -1,0 +1,1 @@
+export const CLOCKWORK_USERAGENT = 'Netlify Clockwork'
