+++ src/utils/functions/get-functions.mjs
@@ -1,17 +1,16 @@
 // @ts-check
-const { fileExistsAsync } = require('../../lib/fs.cjs')
+import { fileExistsAsync } from '../../lib/fs.cjs'
 
 const getUrlPath = (functionName) => `/.netlify/functions/${functionName}`
 
-const BACKGROUND = '-background'
+export const BACKGROUND = '-background'
+const JS = 'js'
 
 const addFunctionProps = ({ mainFile, name, runtime, schedule }) => {
   const urlPath = getUrlPath(name)
   const isBackground = name.endsWith(BACKGROUND)
   return { mainFile, name, runtime, urlPath, isBackground, schedule }
 }
-
-const JS = 'js'
 
 /**
  * @param {Record<string, { schedule?: string }>} functionConfigRecord
@@ -20,7 +19,7 @@ const JS = 'js'
 const extractSchedule = (functionConfigRecord) =>
   Object.fromEntries(Object.entries(functionConfigRecord).map(([name, { schedule }]) => [name, { schedule }]))
 
-const getFunctions = async (functionsSrcDir, config = {}) => {
+export const getFunctions = async (functionsSrcDir, config = {}) => {
   if (!(await fileExistsAsync(functionsSrcDir))) {
     return []
   }
@@ -34,23 +33,3 @@ const getFunctions = async (functionsSrcDir, config = {}) => {
   const functionsWithProps = functions.filter(({ runtime }) => runtime === JS).map((func) => addFunctionProps(func))
   return functionsWithProps
 }
-
-const getFunctionsAndWatchDirs = async (functionsSrcDir) => {
-  if (!(await fileExistsAsync(functionsSrcDir))) {
-    return { functions: [], watchDirs: [functionsSrcDir] }
-  }
-
-  // performance optimization, load '@netlify/zip-it-and-ship-it' on demand
-  const { listFunctions } = await import('@netlify/zip-it-and-ship-it')
-
-  // get all functions files so we know which directories to watch
-  const functions = await listFunctions(functionsSrcDir)
-  const watchDirs = [functionsSrcDir]
-
-  // filter for only main files to serve
-  const functionsWithProps = functions.filter(({ runtime }) => runtime === JS).map((func) => addFunctionProps(func))
-
-  return { functions: functionsWithProps, watchDirs }
-}
-
-module.exports = { getFunctions, getFunctionsAndWatchDirs, BACKGROUND }
