--- src/utils/functions/index.cjs
@@ -1,12 +1,0 @@
-const edgeFunctions = require('../../lib/edge-functions/index.cjs')
-
-const constants = require('./constants.cjs')
-const functions = require('./functions.cjs')
-const getFunctions = require('./get-functions.cjs')
-
-module.exports = {
-  ...constants,
-  ...functions,
-  ...edgeFunctions,
-  ...getFunctions,
-}
